backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.11
alembic==1.13.3
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.2
redis==5.1.1
google-generativeai==0.8.3
openai>=1.54.0
python-multipart==0.0.12
slowapi==0.1.9
